refactor: post, comment를 commnutiy 패키지로 묶기
