+++ growthmate-api/src/main/java/com/growup/growthmate/command/api/CommentV1Controller.java
@@ -3,10 +3,10 @@ package com.growup.growthmate.command.api;
 import com.growup.growthmate.LoginMember;
 import com.growup.growthmate.command.dto.CommentCreateRequest;
 import com.growup.growthmate.command.dto.CommentUpdateRequest;
-import com.growup.growthmate.comment.application.CommentService;
-import com.growup.growthmate.comment.dto.CommentCreateCommand;
-import com.growup.growthmate.comment.dto.CommentDeleteCommand;
-import com.growup.growthmate.comment.dto.CommentUpdateCommand;
+import com.growup.growthmate.community.comment.application.CommentService;
+import com.growup.growthmate.community.comment.dto.CommentCreateCommand;
+import com.growup.growthmate.community.comment.dto.CommentDeleteCommand;
+import com.growup.growthmate.community.comment.dto.CommentUpdateCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
