+++ growthmate-api/src/main/java/com/growup/growthmate/command/api/PostV1Controller.java
@@ -3,10 +3,10 @@ package com.growup.growthmate.command.api;
 import com.growup.growthmate.LoginMember;
 import com.growup.growthmate.command.dto.PostCreateRequest;
 import com.growup.growthmate.command.dto.PostUpdateRequest;
-import com.growup.growthmate.post.application.PostService;
-import com.growup.growthmate.post.dto.PostCreateCommand;
-import com.growup.growthmate.post.dto.PostDeleteCommand;
-import com.growup.growthmate.post.dto.PostUpdateCommand;
+import com.growup.growthmate.community.post.application.PostService;
+import com.growup.growthmate.community.post.dto.PostCreateCommand;
+import com.growup.growthmate.community.post.dto.PostDeleteCommand;
+import com.growup.growthmate.community.post.dto.PostUpdateCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
