--- growthmate-core/src/main/java/com/growup/growthmate/comment/domain/CommentRepository.java
@@ -1,14 +1,0 @@
-package com.growup.growthmate.comment.domain;
-
-import org.springframework.data.repository.Repository;
-
-import java.util.Optional;
-
-public interface CommentRepository extends Repository<Comment, Long> {
-
-    Comment save(Comment comment);
-
-    Optional<Comment> findById(Long id);
-
-    void delete(Comment comment);
-}
