--- growthmate-core/src/main/java/com/growup/growthmate/comment/domain/value/PostId.java
@@ -1,17 +1,0 @@
-package com.growup.growthmate.comment.domain.value;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@EqualsAndHashCode
-@Getter
-public class PostId {
-
-    @Column(nullable = false, name = "post_id")
-    private Long value;
-}
