--- growthmate-core/src/main/java/com/growup/growthmate/comment/dto/CommentCreateCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.comment.dto;
-
-public record CommentCreateCommand(Long postId, String content, Long writerId) {
-}
