--- growthmate-core/src/main/java/com/growup/growthmate/comment/dto/CommentDeleteCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.comment.dto;
-
-public record CommentDeleteCommand(Long commentId, Long writerId) {
-}
