--- growthmate-core/src/main/java/com/growup/growthmate/comment/dto/CommentUpdateCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.comment.dto;
-
-public record CommentUpdateCommand(Long commentId, String content, Long writerId) {
-}
