+++ growthmate-core/src/main/java/com/growup/growthmate/community/WriterId.java
@@ -1,0 +1,16 @@
+package com.growup.growthmate.community;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
+import lombok.*;
+
+@Embeddable
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@EqualsAndHashCode
+@Getter
+public class WriterId {
+
+    @Column(nullable = false, name = "member_id")
+    private Long value;
+}
