+++ growthmate-core/src/main/java/com/growup/growthmate/community/comment/dto/CommentCreateCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.comment.dto;
+
+public record CommentCreateCommand(Long postId, String content, Long writerId) {
+}
