+++ growthmate-core/src/main/java/com/growup/growthmate/community/comment/dto/CommentDeleteCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.comment.dto;
+
+public record CommentDeleteCommand(Long commentId, Long writerId) {
+}
