+++ growthmate-core/src/main/java/com/growup/growthmate/community/comment/dto/CommentUpdateCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.comment.dto;
+
+public record CommentUpdateCommand(Long commentId, String content, Long writerId) {
+}
