+++ growthmate-core/src/main/java/com/growup/growthmate/community/post/dto/PostCreateCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.post.dto;
+
+public record PostCreateCommand(Long companyId, Long writerId, String title, String content) {
+}
