+++ growthmate-core/src/main/java/com/growup/growthmate/community/post/dto/PostDeleteCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.post.dto;
+
+public record PostDeleteCommand(Long postId, Long writerId) {
+}
