+++ growthmate-core/src/main/java/com/growup/growthmate/community/post/dto/PostUpdateCommand.java
@@ -1,0 +1,4 @@
+package com.growup.growthmate.community.post.dto;
+
+public record PostUpdateCommand(Long postId, Long writerId, String title, String content) {
+}
