--- growthmate-core/src/main/java/com/growup/growthmate/post/domain/value/CompanyId.java
@@ -1,16 +1,0 @@
-package com.growup.growthmate.post.domain.value;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@EqualsAndHashCode
-@Getter
-public class CompanyId {
-
-    @Column(nullable = false, name = "company_id")
-    private Long value;
-}
