--- growthmate-core/src/main/java/com/growup/growthmate/post/domain/value/WriterId.java
@@ -1,16 +1,0 @@
-package com.growup.growthmate.post.domain.value;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@EqualsAndHashCode
-@Getter
-public class WriterId {
-
-    @Column(nullable = false, name = "member_id")
-    private Long value;
-}
