--- growthmate-core/src/main/java/com/growup/growthmate/post/dto/PostCreateCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.post.dto;
-
-public record PostCreateCommand(Long companyId, Long writerId, String title, String content) {
-}
