--- growthmate-core/src/main/java/com/growup/growthmate/post/dto/PostDeleteCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.post.dto;
-
-public record PostDeleteCommand(Long postId, Long writerId) {
-}
