--- growthmate-core/src/main/java/com/growup/growthmate/post/dto/PostUpdateCommand.java
@@ -1,4 +1,0 @@
-package com.growup.growthmate.post.dto;
-
-public record PostUpdateCommand(Long postId, Long writerId, String title, String content) {
-}
